move previewWindow class into new file
and cache frequently-created blank frames
